Split the C API methods for TfLiteRegistrationExternal into a separate a separate source file registration_external.cc that is included in a separate build target ":registration_external".  The aim of this change is to ensure proper layering:
   C API ->
   C++ API ->
   C API for Delegates and Custom Ops (i.e. {common,registration_external})

This is needed to allow the C++ API code to reference registration_external
without creating a dependency cycle between the C API and the C++ API.

Right now, TfLiteRegistration is part of "common", but TfliteRegistrationExternal
is part of "c_api", which prevents any references to TfLiteRegistrationExternal
access methods from the C++ API implementation.

PiperOrigin-RevId: 532753083
