+++ tensorflow/lite/core/c/BUILD
@@ -25,6 +25,7 @@ exports_files(
         "c_api_opaque.h",
         "c_api_types.h",
         "common.h",
+        "registration_external.h",
     ],
     visibility = [
         "//tensorflow/lite:__subpackages__",
@@ -43,6 +44,7 @@ filegroup(
         "c_api.h",
         "c_api_types.h",
         "common.h",
+        "registration_external.h",
     ],
 )
 
@@ -59,6 +61,7 @@ tflite_cc_library_with_c_headers_test(
     deps = [
         ":c_api_types",
         ":c_api_without_op_resolver",
+        ":registration_external",
         "//tensorflow/lite:builtin_ops",
         "//tensorflow/lite:create_op_resolver_with_builtin_ops",
         "//tensorflow/lite/c:common",
@@ -91,6 +94,7 @@ tflite_cc_library_with_c_headers_test(
     ],
     deps = [
         ":c_api_types",
+        ":registration_external",
         "//tensorflow/lite:builtin_ops",
         "//tensorflow/lite:version",
         "//tensorflow/lite/c:c_api_internal",
@@ -118,6 +122,7 @@ tflite_cc_library_with_c_headers_test(
     ],
     deps = [
         ":c_api_types",
+        ":registration_external",
         "//tensorflow/lite:builtin_ops",
         "//tensorflow/lite/core/async/c:types",
     ],
@@ -132,6 +137,7 @@ tflite_cc_library_with_c_headers_test(
     tags = ["allow_undefined_symbols"],  # For tflite::CreateOpResolver().
     deps = [
         ":c_api_types",
+        ":registration_external_without_alwayslink",
         "//tensorflow/lite:builtin_ops",
         "//tensorflow/lite:framework",
         "//tensorflow/lite:version",
@@ -398,5 +404,41 @@ cc_test(
         "//tensorflow/lite/delegates:delegate_test_util",
         "//tensorflow/lite/testing:util",
         "@com_google_googletest//:gtest_main",
+    ],
+)
+
+tflite_cc_library_with_c_headers_test(
+    name = "registration_external",
+    srcs = ["registration_external.cc"],
+    hdrs = ["registration_external.h"],
+    compatible_with = get_compatible_with_portable(),
+    copts = tflite_copts(),
+    visibility = [
+        "//tensorflow/lite:__subpackages__",
+    ] + c_api_visibility_allowlist(),
+    deps = [
+        "//tensorflow/lite:builtin_ops",
+        "//tensorflow/lite/c:common_internal",
+        "//tensorflow/lite/core/async/c:types",
+        "//tensorflow/lite/core/c:c_api_types",
+        "//tensorflow/lite/core/c:common",
+    ],
+    # When building shared libraries, we need alwayslink here, since the functions
+    # in this header are part of the public API.
+    alwayslink = 1,
+)
+
+tflite_cc_library_with_c_headers_test(
+    name = "registration_external_without_alwayslink",
+    srcs = ["registration_external.cc"],
+    hdrs = ["registration_external.h"],
+    compatible_with = get_compatible_with_portable(),
+    copts = tflite_copts(),
+    deps = [
+        "//tensorflow/lite:builtin_ops",
+        "//tensorflow/lite/c:common_internal",
+        "//tensorflow/lite/core/async/c:types",
+        "//tensorflow/lite/core/c:c_api_types",
+        "//tensorflow/lite/core/c:common",
     ],
 )
