+++ tensorflow/lite/core/c/c_api.cc
@@ -271,74 +271,6 @@ TfLiteStatus TfLiteTensorCopyToBuffer(const TfLiteTensor* tensor,
   return kTfLiteOk;
 }
 
-TfLiteRegistrationExternal* TfLiteRegistrationExternalCreate(
-    TfLiteBuiltinOperator builtin_code, const char* custom_name, int version) {
-  return new TfLiteRegistrationExternal{/*.custom_name =*/custom_name,
-                                        /*.version =*/version,
-                                        /*.init =*/nullptr,
-                                        /*.free =*/nullptr,
-                                        /*.prepare =*/nullptr,
-                                        /*.invoke =*/nullptr,
-                                        /*.async_kernel =*/nullptr,
-                                        /*.builtin_code =*/builtin_code,
-                                        /*.node_index =*/-1};
-}
-
-void TfLiteRegistrationExternalDelete(TfLiteRegistrationExternal* reg) {
-  delete reg;
-}
-
-void TfLiteRegistrationExternalSetInit(
-    TfLiteRegistrationExternal* registration,
-    void* (*init)(TfLiteOpaqueContext* context, const char* buffer,
-                  size_t length)) {
-  registration->init = init;
-}
-
-void TfLiteRegistrationExternalSetFree(
-    TfLiteRegistrationExternal* registration,
-    void (*free)(TfLiteOpaqueContext* context, void* data)) {
-  registration->free = free;
-}
-
-void TfLiteRegistrationExternalSetPrepare(
-    TfLiteRegistrationExternal* registration,
-    TfLiteStatus (*prepare)(TfLiteOpaqueContext* context,
-                            TfLiteOpaqueNode* node)) {
-  registration->prepare = prepare;
-}
-
-void TfLiteRegistrationExternalSetInvoke(
-    TfLiteRegistrationExternal* registration,
-    TfLiteStatus (*invoke)(TfLiteOpaqueContext* context,
-                           TfLiteOpaqueNode* node)) {
-  registration->invoke = invoke;
-}
-
-void TfLiteRegistrationExternalSetAsyncKernel(
-    TfLiteRegistrationExternal* registration,
-    TfLiteAsyncKernel* (*async_kernel)(TfLiteOpaqueContext* context,
-                                       TfLiteOpaqueNode* node)) {
-  registration->async_kernel = async_kernel;
-}
-
-TfLiteBuiltinOperator TfLiteRegistrationExternalGetBuiltInCode(
-    const TfLiteRegistrationExternal* registration) {
-  return static_cast<TfLiteBuiltinOperator>(registration->builtin_code);
-}
-
-int TfLiteRegistrationExternalGetVersion(
-    const TfLiteRegistrationExternal* registration) {
-  if (!registration) {
-    return -1;
-  }
-  return registration->version;
-}
-
-const char* TfLiteRegistrationExternalGetCustomName(
-    const TfLiteRegistrationExternal* registration) {
-  return registration->custom_name;
-}
 // LINT.ThenChange(//tensorflow/lite/experimental/examples/unity/TensorFlowLitePlugin/Assets/TensorFlowLite/SDK/Scripts/Interpreter.cs)
 
 }  // extern "C"
