+++ tensorflow/lite/core/c/c_api.h
@@ -27,6 +27,7 @@ limitations under the License.
 #include "tensorflow/lite/builtin_ops.h"
 #include "tensorflow/lite/core/async/c/types.h"
 #include "tensorflow/lite/core/c/c_api_types.h"  // IWYU pragma: export
+#include "tensorflow/lite/core/c/registration_external.h"  // IWYU pragma: export
 
 // --------------------------------------------------------------------------
 /// \file
@@ -100,11 +101,6 @@ typedef struct TfLiteInterpreter TfLiteInterpreter;
 /// A tensor in the interpreter system which is a wrapper around a buffer of
 /// data including a dimensionality (or NULL if not currently defined).
 typedef struct TfLiteTensor TfLiteTensor;
-
-/// TfLiteRegistrationExternal is an external version of TfLiteRegistration to
-/// use custom op registration API.
-/// \warning This is an experimental type and subject to change.
-typedef struct TfLiteRegistrationExternal TfLiteRegistrationExternal;
 
 // --------------------------------------------------------------------------
 /// The TensorFlow Lite Runtime version.
@@ -466,96 +462,6 @@ TFL_CAPI_EXPORT extern TfLiteStatus TfLiteTensorCopyFromBuffer(
 TFL_CAPI_EXPORT extern TfLiteStatus TfLiteTensorCopyToBuffer(
     const TfLiteTensor* output_tensor, void* output_data,
     size_t output_data_size);
-
-/// Returns a new TfLiteRegistrationExternal instance.
-///
-/// \note The caller retains ownership and should ensure that
-/// the lifetime of the `TfLiteRegistrationExternal` must be at least as long as
-/// the lifetime of the `TfLiteInterpreter`.
-/// \warning This is an experimental API and subject to change.
-TFL_CAPI_EXPORT extern TfLiteRegistrationExternal*
-TfLiteRegistrationExternalCreate(TfLiteBuiltinOperator builtin_code,
-                                 const char* custom_name, int version);
-
-/// Return the builtin op code of the provided external 'registration'.
-///
-/// \warning This is an experimental API and subject to change.
-TFL_CAPI_EXPORT extern TfLiteBuiltinOperator
-TfLiteRegistrationExternalGetBuiltInCode(
-    const TfLiteRegistrationExternal* registration);
-
-/// Return the OP version of the provided external 'registration'.  Return -1
-/// in case of error, or if the provided address is null.
-///
-/// \warning This is an experimental API and subject to change.
-TFL_CAPI_EXPORT extern int TfLiteRegistrationExternalGetVersion(
-    const TfLiteRegistrationExternal* registration);
-
-/// Returns the custom name of the provided 'registration'. The returned pointer
-/// will be non-null iff the op is a custom op.
-///
-/// \warning This is an experimental API and subject to change.
-TFL_CAPI_EXPORT extern const char* TfLiteRegistrationExternalGetCustomName(
-    const TfLiteRegistrationExternal* registration);
-
-/// Destroys the TfLiteRegistrationExternal instance.
-/// \warning This is an experimental API and subject to change.
-TFL_CAPI_EXPORT extern void TfLiteRegistrationExternalDelete(
-    TfLiteRegistrationExternal* registration);
-
-/// Sets the initialization callback for the registration.
-///
-/// The callback is called to initialize the op from serialized data.
-/// Please refer `init` of `TfLiteRegistration` for the detail.
-/// \warning This is an experimental API and subject to change.
-TFL_CAPI_EXPORT extern void TfLiteRegistrationExternalSetInit(
-    TfLiteRegistrationExternal* registration,
-    void* (*init)(TfLiteOpaqueContext* context, const char* buffer,
-                  size_t length));
-
-/// Sets the deallocation callback for the registration.
-///
-/// This callback is called to deallocate the data returned by the init
-/// callback. The value passed in the `data` parameter is the value that was
-/// returned by the `init` callback.
-/// Please refer `free` of `TfLiteRegistration` for the detail.
-/// \warning This is an experimental API and subject to change.
-TFL_CAPI_EXPORT extern void TfLiteRegistrationExternalSetFree(
-    TfLiteRegistrationExternal* registration,
-    void (*free)(TfLiteOpaqueContext* context, void* data));
-
-/// Sets the preparation callback for the registration.
-///
-/// The callback is called when the inputs of operator have been resized.
-/// Please refer `prepare` of `TfLiteRegistration` for the detail.
-/// \warning This is an experimental API and subject to change.
-TFL_CAPI_EXPORT extern void TfLiteRegistrationExternalSetPrepare(
-    TfLiteRegistrationExternal* registration,
-    TfLiteStatus (*prepare)(TfLiteOpaqueContext* context,
-                            TfLiteOpaqueNode* node));
-
-/// Sets the invocation callback for the registration.
-///
-/// The callback is called when the operator is executed.
-/// Please refer `invoke` of `TfLiteRegistration` for the detail.
-/// \warning This is an experimental API and subject to change.
-TFL_CAPI_EXPORT extern void TfLiteRegistrationExternalSetInvoke(
-    TfLiteRegistrationExternal* registration,
-    TfLiteStatus (*invoke)(TfLiteOpaqueContext* context,
-                           TfLiteOpaqueNode* node));
-
-/// Sets the async kernel accessor callback for the registration.
-///
-/// The callback is called to retrieve the async kernel if the delegate supports
-/// it. If the delegate does not support async execution, either this function
-/// should not be called, or `async_kernel` needs to be nullptr.
-/// `node` is the delegate TfLiteNode created by `ModifyGraphWithDelegate`.
-/// Please refer `async_kernel` of `TfLiteRegistration` for the detail.
-/// \warning This is an experimental API and subject to change.
-TFL_CAPI_EXPORT extern void TfLiteRegistrationExternalSetAsyncKernel(
-    TfLiteRegistrationExternal* registration,
-    TfLiteAsyncKernel* (*async_kernel)(TfLiteOpaqueContext* context,
-                                       TfLiteOpaqueNode* node));
 
 // NOLINTEND(modernize-redundant-void-arg)
