+++ tensorflow/lite/ios/BUILD.apple
@@ -81,9 +81,10 @@ strip_common_include_path_prefix(
         "//tensorflow/lite:builtin_ops.h",
         # Here we include the paths of the actual c library headers in tensorflow/lite/core/c instead of the shim headers in tensorflow/lite/c.
         "//tensorflow/lite/core/c:c_api_experimental.h",
-        "//tensorflow/lite/core/c:common.h",
         "//tensorflow/lite/core/c:c_api.h",
         "//tensorflow/lite/core/c:c_api_types.h",
+        "//tensorflow/lite/core/c:common.h",
+        "//tensorflow/lite/core/c:registration_external.h",
         "//tensorflow/lite/core/async/c:types.h",
         "//tensorflow/lite/delegates/xnnpack:xnnpack_delegate.h",
         "//tensorflow/lite/profiling/telemetry/c:profiler.h",
@@ -214,6 +215,7 @@ cc_library(
         "//tensorflow/lite/core/c:c_api_experimental.h",
         "//tensorflow/lite/core/c:c_api_types.h",
         "//tensorflow/lite/core/c:common.h",
+        "//tensorflow/lite/core/c:registration_external.h",
         "//tensorflow/lite/delegates/xnnpack:xnnpack_delegate.h",
         "//tensorflow/lite/profiling/telemetry/c:profiler.h",
         "//tensorflow/lite/profiling/telemetry/c:telemetry_setting.h",
